Move serialization specific BitArray helpers to where they are used
